refactor : move object projects into separated json file projects.json
then make ajax request to the json file, add onload event on ajax request then get the responseText and parse it into object projects then load the projects as before
this is for implementing JS AJAX to comunicate with back end. We ask the proects list to backend with ajax, then load it.